Package refactor and add new event for Upload Approved Orders
