+++ src/main/java/uk/gov/hmcts/reform/finrem/caseorchestration/handler/uploadapprovedorder/contested/UploadApprovedOrderContestedAboutToSubmitHandler.java
@@ -1,4 +1,4 @@
-package uk.gov.hmcts.reform.finrem.caseorchestration.handler.uploadapprovedorder;
+package uk.gov.hmcts.reform.finrem.caseorchestration.handler.uploadapprovedorder.contested;
 
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
